+++ evolve/template/constants.go
@@ -5,19 +5,18 @@ const (
 
 	RollkitV0XStartHandler = "rollserv.StartHandler"
 	EvolveV1XStartHandler  = "abciserver.StartHandler"
-	evolveV1MigrateCmd     = "abciserver.MigrateToEvolveCmd()"
 )
 
 const (
 	EvABCIPackage = "github.com/evstack/ev-abci"
-	EvABCIVersion = "v1.0.0-beta.3"
+	EvABCIVersion = "v1.0.0-beta.3.0.20251217135843-7ae596a89474" // TODO(@julienrbrt): latest main: bump after tag ev-abci
 
 	EvNodePackage = "github.com/evstack/ev-node"
-	EvNodeVersion = "v1.0.0-beta.7"
+	EvNodeVersion = "v1.0.0-beta.10.0.20251216132820-afcd6bd9b354" // TODO(@julienrbrt): latest main: bump after tag ev-node
 
 	GoHeaderPackageFork = "github.com/julienrbrt/go-header"
 	GoHeaderVersionFork = "v0.0.0-20250909151551-cb11b091bf58"
 	GoHeaderPackage     = "github.com/celestiaorg/go-header"
 
-	EvNodeDaCmd = "github.com/evstack/ev-node/da/cmd/local-da"
+	EvNodeDaCmd = "github.com/evstack/ev-node/tools/local-da"
 )